simplified io module
