--- peepingtom/io_/read/box/__init__.py
@@ -1,1 +1,0 @@
-from .box import read_box
