--- peepingtom/io_/read/em/__init__.py
@@ -1,1 +1,0 @@
-from .em import read_em
