--- peepingtom/io_/read/mrc/__init__.py
@@ -1,1 +1,0 @@
-from .mrc import read_mrc
