--- peepingtom/io_/read/star/__init__.py
@@ -1,1 +1,0 @@
-from .star import read_star
