--- peepingtom/io_/read/tbl/__init__.py
@@ -1,1 +1,0 @@
-from .tbl import read_tbl
